forge fmt

+++ src/oracle/pendle/PendleUniversalAdapter.sol
@@ -19,7 +19,6 @@ import {IERC20} from "forge-std/interfaces/IERC20.sol";
  * @dev Allows fetching oracle data without hardcoding base/quote assets and other parameters
  */
 contract PendleUniversalAdapter {
-
     string public constant name = "PendleUniversalAdapter";
     /// @dev The minimum length of the TWAP window.
     uint32 internal constant MIN_TWAP_WINDOW = 5 minutes;
@@ -29,17 +28,17 @@ contract PendleUniversalAdapter {
     uint8 internal constant FEED_DECIMALS = 18;
     /// @notice The address of the Pendle oracle.
     address public immutable pendleOracle;
-    
+
     /**
      * @notice Constructor for PendleUniversalOracleAdapter
      * @param _pendleOracle Pendle oracle address
      */
     constructor(address _pendleOracle) {
         require(_pendleOracle != address(0), "Invalid Pendle oracle");
-        
+
         pendleOracle = _pendleOracle;
     }
-    
+
     /**
      * @notice Get the price of base asset in quote asset
      * @param base Base asset address
@@ -47,10 +46,14 @@ contract PendleUniversalAdapter {
      * @param twapWindow TWAP window in seconds
      * @return price The price with 18 decimals
      */
-    function getPrice(address pendleMarket, address base, address quote, uint32 twapWindow) external view returns (uint256 price) {
+    function getPrice(address pendleMarket, address base, address quote, uint32 twapWindow)
+        external
+        view
+        returns (uint256 price)
+    {
         price = _getQuote(1e18, pendleMarket, base, quote, twapWindow);
     }
-    
+
     /**
      * @notice Get the quote amount for a given base amount
      * @param inAmount Amount of base asset (with 18 decimals)
@@ -60,10 +63,14 @@ contract PendleUniversalAdapter {
      * @param twapWindow TWAP window in seconds
      * @return quoteAmount The quote amount with 18 decimals
      */
-    function getQuote(uint256 inAmount, address pendleMarket, address base, address quote, uint32 twapWindow) external view returns (uint256 quoteAmount) {
+    function getQuote(uint256 inAmount, address pendleMarket, address base, address quote, uint32 twapWindow)
+        external
+        view
+        returns (uint256 quoteAmount)
+    {
         return _getQuote(inAmount, pendleMarket, base, quote, twapWindow);
     }
-    
+
     /**
      * @notice Get the decimals for the price
      * @return decimals The number of decimal places (18 for Pendle oracles)
@@ -71,7 +78,7 @@ contract PendleUniversalAdapter {
     function decimals() external pure returns (uint8) {
         return 18;
     }
-    
+
     /**
      * @notice Get the description of the oracle
      * @param base Base asset address
@@ -81,7 +88,7 @@ contract PendleUniversalAdapter {
     function description(address base, address quote) external pure returns (string memory) {
         return string(abi.encodePacked("Pendle Oracle: ", _addressToString(base), "/", _addressToString(quote)));
     }
-    
+
     /**
      * @notice Get the version of the oracle
      * @return version The oracle version
@@ -89,7 +96,7 @@ contract PendleUniversalAdapter {
     function version() external pure returns (uint256) {
         return 1;
     }
-    
+
     /**
      * @notice Get the quote amount for a given base amount (internal function)
      * @param inAmount The amount of `base` to convert.
@@ -98,7 +105,11 @@ contract PendleUniversalAdapter {
      * @param twapWindow TWAP window in seconds
      * @return quoteAmount The quote amount
      */
-    function _getQuote(uint256 inAmount, address pendleMarket, address base, address quote, uint32 twapWindow) internal view returns (uint256 quoteAmount) {
+    function _getQuote(uint256 inAmount, address pendleMarket, address base, address quote, uint32 twapWindow)
+        internal
+        view
+        returns (uint256 quoteAmount)
+    {
         _validateTwapWindow(twapWindow, pendleOracle, pendleMarket);
 
         (address _base, address _quote) = _getBaseAndQuote(pendleMarket, base, quote);
@@ -106,7 +117,7 @@ contract PendleUniversalAdapter {
         uint256 unitPrice = _getRateFunction(pendleMarket, _base, _quote)(IPMarket(pendleMarket), twapWindow);
         return ScaleUtils.calcOutAmount(inAmount, unitPrice, _getScale(_base, _quote), inverse);
     }
-    
+
     /**
      * @notice Convert address to string
      * @param addr Address to convert
@@ -114,8 +125,8 @@ contract PendleUniversalAdapter {
      */
     function _addressToString(address addr) internal pure returns (string memory) {
         bytes memory b = new bytes(20);
-        for (uint i = 0; i < 20; i++) {
-            b[i] = bytes1(uint8(uint160(addr) / (2**(8*(19 - i)))));
+        for (uint256 i = 0; i < 20; i++) {
+            b[i] = bytes1(uint8(uint160(addr) / (2 ** (8 * (19 - i)))));
         }
         return string(b);
     }
@@ -133,7 +144,11 @@ contract PendleUniversalAdapter {
         }
     }
 
-    function _getRateFunction(address pendleMarket, address base, address quote) internal view returns (function(IPMarket, uint32) view returns (uint256)) {
+    function _getRateFunction(address pendleMarket, address base, address quote)
+        internal
+        view
+        returns (function(IPMarket, uint32) view returns (uint256))
+    {
         (IStandardizedYield sy, IPPrincipalToken pt,) = IPMarket(pendleMarket).readTokens();
         if (base == address(pt)) {
             if (quote == address(sy)) {
@@ -158,7 +173,11 @@ contract PendleUniversalAdapter {
         return ScaleUtils.calcScale(baseDecimals, quoteDecimals, FEED_DECIMALS);
     }
 
-    function _getBaseAndQuote(address pendleMarket, address base, address quote) internal view returns (address, address) {
+    function _getBaseAndQuote(address pendleMarket, address base, address quote)
+        internal
+        view
+        returns (address, address)
+    {
         (IStandardizedYield sy, IPPrincipalToken pt,) = IPMarket(pendleMarket).readTokens();
         if (base == address(pt)) {
             if (quote == address(sy)) {
